aw-sync: Moved into a seperate member
