+++ aw-sync/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "aw-sync"
+version = "0.1.0"
+authors = ["Johan Bjäreholt <johan@bjareho.lt>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+log = "0.4"
+chrono = { version = "0.4", features = ["serde"] }
+serde = "1.0"
+serde_json = "1.0"
+aw_server = { path = "../" }
+aw-models = { path = "../aw-models" }
+aw-datastore = { path = "../aw-datastore" }
